# apps/mrerp/backend/openapi.yaml
openapi: 3.0.3
info:
  title: MRERP API
  version: 1.0.0
  description: MRERP modular-monolith API contract
paths:
...
  /api/v1/leave/actual-attendance/:
    get:
      operationId: api_v1_leave_actual_attendance_list
      parameters:
      - in: query
        name: month
        schema:
          type: string
        required: true
      tags:
      - api
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AttendanceResult'
          description: ''
  /api/v1/leave/attendance/:
    get:
      operationId: api_v1_leave_attendance_list
      parameters:
      - in: query
        name: month
        schema:
          type: string
        description: Tháng định dạng YYYY-MM.
        required: true
      tags:
      - api
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AttendanceRow'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: ''
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        code:
          type: string
        detail:
          type: string
        correlation_id:
          type: string
          format: uuid
        errors:
          type: object
          additionalProperties: {}
      required:
      - code
      - correlation_id
      - detail
...
    AttendanceResult:
      type: object
      properties:
        month:
          type: string
        employees:
          type: array
          items:
            type: object
            additionalProperties: {}
      required:
      - employees
      - month
    AttendanceRow:
      type: object
      properties:
        employee_uuid:
          type: string
          format: uuid
        employee_code:
          type: string
        display_name:
          type: string
        team_name:
          type: string
          nullable: true
        month:
          type: string
        scheduled_workdays:
          type: number
          format: double
        public_holiday_days:
          type: number
          format: double
        approved_leave_days:
          type: number
          format: double
        adjustment_days:
          type: integer
        adjustment_reason:
          type: string
        projected_workdays:
          type: number
          format: double
      required:
      - adjustment_days
      - adjustment_reason
      - approved_leave_days
      - display_name
      - employee_code
      - employee_uuid
      - month
      - projected_workdays
      - public_holiday_days
      - scheduled_workdays
      - team_name
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid